Resetting work on Player Sheet + Gear Items.
I wrote a lot of code while in the early stages of recovering from a
surgery, and honestly I hated it. Particularly the DOM, but also the
gear item type ended up incredibly unwieldy.

Clearing these out and getting ready for a second take on all of it.
Each sub-type of gear will now have a dedicated item type and its own
sheet class, both of which will be derived from a base, abstract Gear
model/sheet.